Mon, Jun 13, 2022, 5:20 AM +00:00

+++ bitcoin/bitcoin.py
@@ -12,7 +12,7 @@ if len(sys) == 2:
             y = x.json()
             rate = y["bpi"]['USD']['rate']
             newrate = rate.replace(',','')
-            total = float((float(newrate))*int(sys[1]))
+            total = float((float(newrate))*float(sys[1]))
             print("$"+"{:,}".format(total))
